+++ Content.Shared/Random/Helpers/SharedRandomExtensions.cs
@@ -109,7 +109,7 @@ namespace Content.Shared.Random.Helpers
             var sum = picks.Values.Sum();
             var accumulated = 0f;
 
-            var rand = random.NextFloat() * sum;
+            var rand = (float) (random.NextDouble() * sum);
 
             foreach (var (key, weight) in picks)
             {
@@ -195,7 +195,7 @@ namespace Content.Shared.Random.Helpers
             var sum = weights.Values.Sum();
             var accumulated = 0f;
 
-            var rand = random.NextFloat() * sum;
+            var rand = (float) (random.NextDouble() * sum);
 
             foreach (var (key, weight) in weights)
             {
